Added: ChatManager.
