+++ src/main/java/dev/boostio/events/AsyncPlayerPreLogin.java
@@ -15,7 +15,7 @@ import java.util.UUID;
 public class AsyncPlayerPreLogin implements Listener {
     private final ConfigManager configManager;
 
-    public AsyncPlayerPreLogin(ChatPro plugin){
+    public AsyncPlayerPreLogin(ChatPro plugin) {
         this.configManager = plugin.getConfigManager();
     }
 
@@ -26,13 +26,10 @@ public class AsyncPlayerPreLogin implements Listener {
         PlayerData playerData = new PlayerData();
         ChatPro.getInstance().getPlayerData().put(uuid, playerData);
 
-        ChatColor defaultColor = ColoringUtils.convertColor(getDefaultChatColor());
+        ChatColor defaultColor = ColoringUtils.convertColor(configManager.getConfigurationOption(ConfigOption.DEFAULT_CHAT_COLOR));
 
         //Set default chat color.
         ChatPro.getInstance().getPlayerData().get(uuid).setChatColorName(defaultColor);
     }
-
-    private String getDefaultChatColor(){
-        return configManager.getConfigurationOption(ConfigOption.DEFAULT_CHAT_COLOR);
-    }
 }
+
